requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.0.1

# WebSockets
websockets==12.0

# Selenium para scraping
selenium==4.15.0

# Utilidades
python-dotenv==1.0.0

# Base de datos (opcional para después)
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Cache (opcional para después)
redis==5.0.1

# Tareas asíncronas (opcional para después)
celery==5.3.4

# Desarrollo
pytest==7.4.3
pytest-asyncio==0.21.1

django>=5.0.0

aiohttp==3.11.10
requests==2.32.3

brotli==1.1.0

alembic==1.13.0

python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.6
python-decouple==3.8
pydantic[email]==2.5.0

